fixed Wextra error, changed int to size_t for comparison

+++ 0x1D-search_algorithms/0-linear.c
@@ -9,13 +9,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int index;
+	size_t index;
 
 	if (array == NULL)
 		return (-1);
 	for (index = 0; index < size; index++)
 	{
-		printf("Value checked array[%d] = [%d]\n", index, array[index]);
+		printf("Value checked array[%lu] = [%d]\n",
+		       index, array[index]);
 		if (value == array[index])
 			return (index);
 	}
